+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - It's now possible to add new group members by their email addresses
+- Experimental support for webxdc
 
 ### Changed
 - Update translations (3.1.2022)